% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_detection.R
\name{clean_detection}
\alias{clean_detection}
\title{Clean Cytokine Concentration Data}
\usage{
clean_detection(data)
}
\arguments{
\item{data}{\link{data.frame} of cytokine detection limit}
}
\value{
Returns a \link{data.frame} after initial data cleaning
}
\description{
Clean Cytokine Concentration Data
}
