% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_immune_indicators.R
\name{create_immune_indicators}
\alias{create_immune_indicators}
\title{Create Immune Activation Indicators}
\usage{
create_immune_indicators(
  data,
  immune_vars = c("m_type_1_db", "m_ms", "m_hashimotos", "m_graves",
    "m_rheumatoid_arthritis", "m_psoriasis", "m_rheumatic_carditis", "m_alopecia",
    "m_crohns_disease", "m_ulcerative_colitis", "m_coeliac", "m_lupus_erythematosus",
    "m_sjogrens", "m_antiphospholipid_syndrome", "m_asthma", "m_allergy", "m_eczema",
    "m_idiopathic_thrombocytopenic_purpura", "m_uticarial_vasculitis",
    "m_white_dot_syndrome", "m_inflammatory_bowel_type_unknown", "m_kawasaki",
    "m_raynauds", "m_stevens_johnson", "m_eosinophile_follicueitis", 
    
    "m_other_autoimmune"),
  atopic_vars = c("m_asthma", "m_allergy", "m_eczema")
)
}
\arguments{
\item{data}{\link{data.frame} of the cytokine concentrations}

\item{immune_vars}{A vector of variables used to construct \code{Immune}}

\item{atopic_vars}{A vector of variables used to construct \verb{Atopic Allergy}}
}
\value{
Returns a \link{data.frame} with the immune activation indicators
}
\description{
Create Immune Activation Indicators
}
