% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_cytokines.R
\name{impute_cytokines}
\alias{impute_cytokines}
\title{Impute cytokines}
\usage{
impute_cytokines(
  data,
  cyto_vars = c("sCD40L (12)", "EGF (13)", "Eotaxin (14)", "FGF-2 (15)", "FLT-3L (18)",
    "Fractalkine (19)", "G-CSF (20)", "GM-CSF (21)", "GROa (22)", "IFNa2 (25)",
    "IFNg (26)", "IL-1a (27)", "IL-1b (28)", "IL-1ra (29)", "IL-2 (30)", "IL-3 (33)",
    "IL-4 (34)", "IL-5 (35)", "IL-6 (36)", "IL-7 (37)", "IL-8 (38)", "IL-9 (39)",
    "IL-10 (42)", "IL-12p40 (43)", "IL-12p70 (44)", "IL-13 (46)", "IL-15 (47)",
    "IL-17a (48)", "IL-17e (51)", "IL-17F (53)", "IL-18 (54)", "IL-22 (55)",
    "IL-27 (56)", "IP-10 (57)", "MCP-1 (61)", 
     "MCP-3 (62)", "M-CSF (63)",
    "MDC (64)", "MIG (65)", "MIP-1a (66)", "MIP-1b (67)", "PDGF-AA (72)",
    "PDGF-AB/BB (73)", "TGFa (75)", "TNFa (76)", "TNFb (77)", "VEGF-A (78)"),
  threshold = 0.55,
  my_seed = 649
)
}
\arguments{
\item{data}{A \link{data.frame} that includes cytokine concentrations}

\item{cyto_vars}{A vector of cytokine variable names}

\item{threshold}{A \link{numeric} value that indicates the level of missingness
to tolerate for imputation. Values above the threshold will be excluded.}

\item{my_seed}{\link{numeric} Set seed for reproducibility}
}
\value{
A \link{data.frame} with imputed cytokine data for all cytokines with
missingness below the threshold. Summary of cytokine missingness can be
found in the 'missing' attribute.
}
\description{
Impute cytokines
}
